fix(agents): align OpenRouter logsRoot between write and read sides
The write side (optionsAdapter → OpenRouterAgentRun.logsRoot) and the read
side (OpenRouterSessionProvider) were resolving the logs root differently:

  - Write side: only set logsRoot if user provided one in Settings; otherwise
    OR's own default `<cwd>/logs` kicked in. That polluted the user's project
    directory AND diverged from the read side's default.
  - Read side: always defaulted to ~/.openrouter-agent-coder/logs.

Result: OR sessions wrote to <project>/logs/<sessionId>/ but
GET /chats/:id/messages looked in ~/.openrouter-agent-coder/logs/ — found
nothing — and the entire chat history endpoint silently returned []. This
also explained "no chat in the sidebar" since the discoverSessions scan
hit an empty (or nonexistent) directory.

Extract the resolver to a single helper `resolveOpenRouterLogsRoot()` and
route both sides through it. Resolution order is unchanged from what the
plan called out:

  1. getAgentSettings().openRouterLogsRoot (user override)
  2. $XDG_DATA_HOME/openrouter-agent-coder/logs
  3. ~/.openrouter-agent-coder/logs (default)

For new sessions, the write side now ALWAYS sets logsRoot — OR's own
`<cwd>/logs` default is never used. Existing on-disk sessions written
under the old default will not be discovered by the SessionProvider
(users with such sessions can either move the dir or set the override
in Settings → API → OpenRouter → Logs Root).

82/82 OR adapter tests pass. Build clean.
